refactor: decompose 3 more monolith files into submodules (issue #2486 sprint 7)
- streaming.py (884 → 54 LOC) → _streaming_config, _streaming_protocol,
  _streaming_buffer, _streaming_server, _simulation_streamer
- c3d_reader.py/bio (879 → 280 LOC) → _c3d_models, _c3d_io, _c3d_markers, _c3d_analog
- simulator.py/putting_green (865 → 33 LOC) → _sim_config, _sim_core,
  _wind_physics, _practice_mode

All public symbols remain importable from original module paths.
