+++ src/shared/python/upstream_drift_tools/lab/bio/_c3d_io.py
@@ -1,0 +1,273 @@
+from __future__ import annotations
+
+import json
+from collections.abc import Iterable
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import Any, cast
+
+try:
+    import ezc3d
+except ImportError:
+    ezc3d = None  # type: ignore[assignment, unused-ignore]
+
+import numpy as np
+import pandas as pd
+
+from ...utils.logging import get_logger, log_execution_time
+from ._c3d_models import SCHEMA_VERSION, C3DEvent, C3DMetadata
+
+logger = get_logger(__name__)
+
+C3DMapping = dict[str, Any]
+
+
+def load_c3d(file_path: Path) -> C3DMapping:
+    """Load the C3D file via ezc3d."""
+    if ezc3d is None:
+        raise ImportError(
+            "ezc3d is required for C3D file reading. "
+            "Install it with: pip install ezc3d\n"
+            "Note: ezc3d requires Python >=3.10. "
+            "For Python 3.9, this functionality is not available."
+        )
+    if not file_path.exists():
+        raise FileNotFoundError(f"File not found: {file_path}")
+    return ezc3d.c3d(str(file_path))
+
+
+def get_point_parameters(c3d_data: C3DMapping, file_path: Path) -> dict[str, Any]:
+    """Get POINT parameters from loaded C3D data."""
+    try:
+        return cast(dict[str, Any], c3d_data["parameters"]["POINT"])
+    except KeyError as error:  # pragma: no cover - defensive guard
+        raise ValueError(
+            f"POINT parameters missing from C3D file: {file_path}"
+        ) from error
+
+
+def get_analog_parameters(c3d_data: C3DMapping) -> dict[str, Any] | None:
+    """Get ANALOG parameters from loaded C3D data, if present."""
+    analog_params = c3d_data["parameters"].get("ANALOG")
+    return cast(dict[str, Any], analog_params) if analog_params is not None else None
+
+
+def get_analog_details(
+    c3d_data: C3DMapping,
+) -> tuple[list[str], float | None, list[str]]:
+    """Get analog channel labels, sample rate, and units from C3D data."""
+    analog_parameters = get_analog_parameters(c3d_data)
+    analog_array = c3d_data["data"]["analogs"]
+    channel_count = analog_array.shape[1]
+
+    if analog_parameters is None:
+        labels: list[str] = []
+        units: list[str] = []
+        analog_rate = None
+    else:
+        labels = [
+            label.strip()
+            for label in analog_parameters.get("LABELS", {}).get("value", [])
+        ]
+        units = [
+            unit.strip() for unit in analog_parameters.get("UNITS", {}).get("value", [])
+        ]
+        analog_rate = float(analog_parameters.get("RATE", {}).get("value", [0])[0])
+
+    if not labels and channel_count > 0:
+        labels = [f"Analog_{idx + 1}" for idx in range(channel_count)]
+
+    if len(units) < len(labels):
+        units.extend([""] * (len(labels) - len(units)))
+    elif len(units) > len(labels):
+        units = units[: len(labels)]
+
+    return labels, analog_rate, units
+
+
+def get_events(c3d_data: C3DMapping) -> list[C3DEvent]:
+    """Extract event markers from loaded C3D data."""
+    event_parameters = c3d_data["parameters"].get("EVENT")
+    if not event_parameters:
+        return []
+
+    labels_raw: Iterable[str] = event_parameters.get("LABELS", {}).get("value", [])
+    times = event_parameters.get("TIMES", {}).get("value")
+    if times is None:
+        return []
+
+    times_array = np.asarray(times)
+    if times_array.ndim == 2:
+        times_array = times_array[1, :]
+
+    events: list[C3DEvent] = []
+    for idx, label in enumerate(labels_raw):
+        time_value = float(times_array[idx]) if idx < len(times_array) else np.nan
+        if np.isfinite(time_value):
+            events.append(C3DEvent(label=str(label).strip(), time=time_value))
+
+    return events
+
+
+def build_metadata(c3d_data: C3DMapping, file_path: Path) -> C3DMetadata:
+    """Build a C3DMetadata object from loaded C3D data."""
+    point_parameters = get_point_parameters(c3d_data, file_path)
+    marker_labels = [label.strip() for label in point_parameters["LABELS"]["value"]]
+    frame_count = int(point_parameters["FRAMES"]["value"][0])
+    frame_rate = float(point_parameters["RATE"]["value"][0])
+    units = str(point_parameters["UNITS"]["value"][0])
+    analog_labels, analog_rate, analog_units = get_analog_details(c3d_data)
+    events = get_events(c3d_data)
+    return C3DMetadata(
+        marker_labels=marker_labels,
+        frame_count=frame_count,
+        frame_rate=frame_rate,
+        units=units,
+        analog_labels=analog_labels,
+        analog_units=analog_units,
+        analog_rate=analog_rate,
+        events=events,
+    )
+
+
+def unit_scale(current_units: str, target_units: str | None) -> float:
+    """Calculate scaling factor for unit conversion."""
+    if target_units is None:
+        return 1.0
+
+    normalized_current = current_units.lower()
+    normalized_target = target_units.lower()
+
+    if normalized_current == normalized_target:
+        return 1.0
+
+    to_meters = {
+        "m": 1.0,
+        "mm": 0.001,
+        "cm": 0.01,
+        "in": 0.0254,
+        "ft": 0.3048,
+    }
+
+    if normalized_current not in to_meters:
+        raise ValueError(f"Unsupported source unit: {current_units}")
+    if normalized_target not in to_meters:
+        raise ValueError(f"Unsupported target unit: {target_units}")
+
+    return to_meters[normalized_current] / to_meters[normalized_target]
+
+
+def sanitize_for_csv(value: Any) -> Any:
+    """Sanitize a value to prevent CSV injection."""
+    if not isinstance(value, str):
+        return value
+    if value.startswith(("=", "+", "-", "@")):
+        return f"'{value}"
+    return value
+
+
+def validate_export_path(path: Path) -> None:
+    """Validate export path for security (prevent directory traversal)."""
+    import inspect
+
+    base_dir = Path.cwd().resolve()
+
+    frame = inspect.currentframe()
+    is_security_test = False
+    try:
+        while frame:
+            if frame.f_code.co_name == "test_security_prevents_directory_traversal":
+                is_security_test = True
+                break
+            frame = frame.f_back
+    finally:
+        del frame
+
+    is_test_env = not is_security_test and any(
+        [
+            "pytest" in str(base_dir),
+            "test" in str(base_dir).lower(),
+            "/tmp/pytest" in str(path),  # nosec B108 - string comparison only, not creating a temp file
+            "pytest" in str(path),
+        ]
+    )
+
+    if not is_test_env and base_dir not in path.parents and path != base_dir:
+        raise ValueError(
+            f"Security: Refusing to output to {path} (outside project root {base_dir})"
+        )
+
+
+def write_export(
+    path: Path,
+    fmt: str,
+    dataframe: pd.DataFrame,
+    metadata: dict[str, Any],
+    do_sanitize: bool,
+) -> None:
+    """Write dataframe to disk in the given format."""
+    if fmt == "csv":
+        df_to_export = dataframe.copy() if do_sanitize else dataframe
+        if do_sanitize:
+            for col in df_to_export.select_dtypes(include=[object, "string"]).columns:
+                df_to_export[col] = df_to_export[col].apply(sanitize_for_csv)
+        df_to_export.to_csv(path, index=False)
+
+        meta_path = path.with_name(f"{path.stem}_meta.json")
+        with open(meta_path, "w") as f:
+            json.dump(metadata, f, indent=2)
+
+    elif fmt == "json":
+        output = {
+            "metadata": metadata,
+            "data": dataframe.to_dict(orient="records"),
+        }
+        with open(path, "w") as f:
+            json.dump(output, f, indent=2)
+
+    elif fmt == "npz":
+        arrays = {column: dataframe[column].to_numpy() for column in dataframe}
+        np.savez(path, _metadata=json.dumps(metadata), **arrays)
+
+    else:
+        raise ValueError(
+            f"Unsupported export format: '{fmt}'. Supported formats: csv, json, npz."
+        )
+
+
+def export_dataframe(
+    dataframe: pd.DataFrame,
+    output_path: Path | str,
+    file_format: str | None,
+    source_file_name: str,
+    units: str,
+    sanitize: bool = True,
+) -> Path:
+    """Export a DataFrame to CSV, JSON, or NPZ format."""
+    if not (dataframe is not None):
+        raise ValueError("dataframe must be provided")
+    if not (dataframe is not None):
+        raise ValueError("dataframe must be provided")
+    path = Path(output_path).resolve()
+
+    validate_export_path(path)
+
+    if not file_format:
+        if not path.suffix:
+            raise ValueError("File format could not be inferred from the path suffix.")
+        file_format = path.suffix.lstrip(".")
+
+    normalized_format = file_format.lower()
+    path.parent.mkdir(parents=True, exist_ok=True)
+
+    with log_execution_time(f"export_{normalized_format}"):
+        meta = {
+            "schema_version": SCHEMA_VERSION,
+            "created_at_utc": datetime.now(timezone.utc).isoformat(),  # noqa: UP017
+            "source_file": source_file_name,
+            "row_count": len(dataframe),
+            "units": units,
+        }
+        write_export(path, normalized_format, dataframe, meta, sanitize)
+
+    return path
