+++ src/shared/python/upstream_drift_tools/lab/bio/c3d_reader.py
@@ -5,97 +5,45 @@ Migrated from Golf_Modeling_Suite.
 
 from __future__ import annotations
 
-import json
-from collections.abc import Iterable, Sequence
-from dataclasses import dataclass
-from datetime import datetime, timezone
+from collections.abc import Sequence
 from pathlib import Path
-from typing import Any, cast
 
-try:
-    import ezc3d
-except ImportError:
-    ezc3d = None  # type: ignore[assignment, unused-ignore]
-
-import numpy as np
 import pandas as pd
 
-from ...utils.logging import get_logger, log_execution_time
+from ...utils.logging import get_logger
+from ._c3d_analog import (
+    build_analog_dataframe,
+    build_force_plate_dataframe,
+    detect_force_plate_channels,
+    force_plate_columns,
+)
+from ._c3d_io import (
+    C3DMapping,
+    build_metadata,
+    export_dataframe,
+    load_c3d,
+    unit_scale,
+)
+from ._c3d_markers import build_points_dataframe
+from ._c3d_models import (
+    BIOMECHANICAL_MARKER_MAX_M,
+    BIOMECHANICAL_MARKER_MIN_M,
+    SCHEMA_VERSION,
+    C3DEvent,
+    C3DMetadata,
+)
 
 logger = get_logger(__name__)
 
-C3DMapping = dict[str, Any]
-SCHEMA_VERSION = "1.0"
-
-# Guideline P1: Biomechanical marker validation thresholds [m]
-# Source: NIST - Human body dimensions range from
-# ~0.001m (1mm detail) to ~10m (extended reach)
-BIOMECHANICAL_MARKER_MIN_M = 0.001  # 1mm minimum - detects mm/m confusion
-BIOMECHANICAL_MARKER_MAX_M = 10.0  # 10m maximum - detects unrealistic scales
-
-
-@dataclass(frozen=True)
-class C3DEvent:
-    """A labeled event occurring at a specific time within a capture."""
-
-    label: str
-    time: float
-
-    def __post_init__(self) -> None:
-        """Validate event data."""
-        if not self.label:
-            raise ValueError("Event label cannot be empty.")
-        # Time can be negative (pre-trigger) per spec, so we allow it.
-
-
-@dataclass(frozen=True)
-class C3DMetadata:
-    """Describes key properties of a C3D motion-capture recording."""
-
-    marker_labels: list[str]
-    frame_count: int
-    frame_rate: float
-    units: str
-    analog_labels: list[str]
-    analog_units: list[str]
-    analog_rate: float | None
-    events: list[C3DEvent]
-
-    def __post_init__(self) -> None:
-        """Validate metadata fields."""
-        if self.frame_count < 0:
-            raise ValueError(f"Frame count cannot be negative: {self.frame_count}")
-        if self.frame_rate < 0:
-            raise ValueError(f"Frame rate cannot be negative: {self.frame_rate}")
-        if self.analog_rate is not None and self.analog_rate < 0:
-            raise ValueError(f"Analog rate cannot be negative: {self.analog_rate}")
-
-        # Check consistency
-        if len(self.analog_units) != len(self.analog_labels):
-            raise ValueError(
-                "analog_units and analog_labels must have the same length: "
-                f"{len(self.analog_units)} units vs {len(self.analog_labels)} labels"
-            )
-
-    @property
-    def marker_count(self) -> int:
-        """Number of tracked markers in the recording."""
-
-        return len(self.marker_labels)
-
-    @property
-    def analog_count(self) -> int:
-        """Number of analog channels in the recording."""
-
-        return len(self.analog_labels)
-
-    @property
-    def duration(self) -> float:
-        """Capture duration in seconds, or ``0`` if the rate is missing."""
-
-        if self.frame_rate == 0:
-            return 0.0
-        return self.frame_count / self.frame_rate
+__all__ = [
+    "C3DDataReader",
+    "C3DEvent",
+    "C3DMapping",
+    "C3DMetadata",
+    "BIOMECHANICAL_MARKER_MAX_M",
+    "BIOMECHANICAL_MARKER_MIN_M",
+    "SCHEMA_VERSION",
+]
 
 
 class C3DDataReader:
@@ -113,28 +61,8 @@ class C3DDataReader:
 
     def get_metadata(self) -> C3DMetadata:
         """Return metadata describing marker labels, frame count, rate, and units."""
-
         if self._metadata is None:
-            point_parameters = self._get_point_parameters()
-            marker_labels = [
-                label.strip() for label in point_parameters["LABELS"]["value"]
-            ]
-            frame_count = int(point_parameters["FRAMES"]["value"][0])
-            frame_rate = float(point_parameters["RATE"]["value"][0])
-            units = str(point_parameters["UNITS"]["value"][0])
-            analog_labels, analog_rate, analog_units = self._get_analog_details()
-            events = self._get_events()
-            self._metadata = C3DMetadata(
-                marker_labels=marker_labels,
-                frame_count=frame_count,
-                frame_rate=frame_rate,
-                units=units,
-                analog_labels=analog_labels,
-                analog_units=analog_units,
-                analog_rate=analog_rate,
-                events=events,
-            )
-
+            self._metadata = build_metadata(self._load(), self.file_path)
         return self._metadata
 
     def points_dataframe(
@@ -163,120 +91,23 @@ class C3DDataReader:
             ``residual`` (EzC3D stores residuals in the fourth point channel), and
             an optional ``time`` column in seconds.
         """
-
         if not (include_time is not None):
             raise ValueError("include_time must be provided")
         if not (include_time is not None):
             raise ValueError("include_time must be provided")
         c3d_data = self._load()
         metadata = self.get_metadata()
-        points = c3d_data["data"]["points"]
-
-        marker_labels = np.array(metadata.marker_labels)
-
-        if markers:
-            mask = np.isin(marker_labels, list(markers))
-            marker_labels = marker_labels[mask]
-            points = points[:, mask, :]
-
-        # Sort markers alphabetically to avoid expensive DataFrame sorting later
-        sort_indices = np.argsort(marker_labels)
-        sorted_labels = marker_labels[sort_indices]
-        points = points[:, sort_indices, :]
-
-        raw_coordinates = np.transpose(points[:3, :, :], axes=(2, 1, 0)).reshape(-1, 3)
-        coordinates = raw_coordinates * self._unit_scale(metadata.units, target_units)
-
-        self._validate_marker_positions(coordinates, metadata.units, target_units)
-
-        residuals = points[3, :, :].T.reshape(-1)
-
-        if residual_nan_threshold is not None:
-            too_noisy = residuals > residual_nan_threshold
-            coordinates[too_noisy, :] = np.nan
-
-        current_marker_count = len(sorted_labels)
-        frame_indices = np.repeat(np.arange(metadata.frame_count), current_marker_count)
-        marker_names = np.tile(sorted_labels, metadata.frame_count)
-
-        data: dict[str, Any] = {
-            "frame": frame_indices,
-            "marker": marker_names,
-            "x": coordinates[:, 0],
-            "y": coordinates[:, 1],
-            "z": coordinates[:, 2],
-            "residual": residuals,
-        }
-
-        if include_time:
-            if metadata.frame_rate > 0:
-                data["time"] = frame_indices / metadata.frame_rate
-            else:
-                logger.warning(
-                    "Frame rate is 0. Time column will be omitted "
-                    "despite include_time=True."
-                )
-
-        dataframe = pd.DataFrame(data)
-        dataframe = dataframe.reset_index(drop=True)
-
-        logger.info(
-            "Loaded %s frames for %s markers from %s",
-            metadata.frame_count,
-            current_marker_count,
+        scale = unit_scale(metadata.units, target_units)
+        return build_points_dataframe(
+            c3d_data,
+            metadata,
             self.file_path.name,
+            scale,
+            include_time,
+            markers,
+            residual_nan_threshold,
+            target_units,
         )
-        return dataframe
-
-    @staticmethod
-    def _validate_marker_positions(
-        coordinates: np.ndarray,
-        source_units: str,
-        target_units: str | None,
-    ) -> None:
-        """Validate marker positions per Guideline P1 (biomechanical range check).
-
-        Raises:
-            ValueError: If positions exceed the 10m sanity threshold.
-        """
-        if coordinates.size == 0:
-            return
-
-        min_pos = np.nanmin(coordinates)
-        max_pos = np.nanmax(coordinates)
-
-        if np.isnan(min_pos) or np.isnan(max_pos):
-            logger.warning(
-                "All marker coordinates are NaN or non-finite; skipping unit "
-                "range validation (Guideline P1). Verify upstream data quality "
-                "and missing-data handling."
-            )
-            return
-
-        if min_pos < BIOMECHANICAL_MARKER_MIN_M:
-            logger.warning(
-                "⚠️ Suspiciously small marker positions detected (< 1mm). "
-                f"Min position: {min_pos:.6f}m. "
-                f"Source units: {source_units}, target: "
-                f"{target_units or 'unchanged'}. "
-                "Guideline P1: Verify unit conversion is correct to "
-                "avoid 1000x errors."
-            )
-
-        if max_pos > BIOMECHANICAL_MARKER_MAX_M:
-            logger.error(
-                "❌ Unrealistic marker positions detected (> 10m). "
-                f"Max position: {max_pos:.2f}m. "
-                f"Source units: {source_units}, target: "
-                f"{target_units or 'unchanged'}. "
-                "Guideline P1 VIOLATION: Likely unit conversion error."
-            )
-            raise ValueError(
-                f"Marker positions exceed {BIOMECHANICAL_MARKER_MAX_M}m "
-                f"(max: {max_pos:.2f}m) - likely unit error. "
-                f"Check that source units '{source_units}' are correct. "
-                "Common issue: mm labeled as m or vice versa."
-            )
 
     def analog_dataframe(self, include_time: bool = True) -> pd.DataFrame:
         """Return analog channels as a tidy DataFrame.
@@ -284,45 +115,11 @@ class C3DDataReader:
         Rows are ordered by sample index and channel name so downstream GUI
         components can easily plot synchronized sensor traces.
         """
-
         if not (include_time is not None):
             raise ValueError("include_time must be provided")
         if not (include_time is not None):
             raise ValueError("include_time must be provided")
-        c3d_data = self._load()
-        metadata = self.get_metadata()
-        analog_array = c3d_data["data"]["analogs"]
-        subframes, channel_count, frame_count = analog_array.shape
-        analog_rate = metadata.analog_rate
-
-        columns = ["sample", "channel", "value"]
-        if include_time and analog_rate:
-            columns = ["sample", "time", "channel", "value"]
-
-        if channel_count == 0:
-            return pd.DataFrame(columns=columns)
-
-        values = analog_array.transpose(2, 0, 1).reshape(
-            frame_count * subframes, channel_count
-        )
-        sample_indices = np.arange(values.shape[0])
-        channel_names = np.array(
-            metadata.analog_labels
-            or [f"Analog_{idx + 1}" for idx in range(channel_count)]
-        )
-
-        dataframe = pd.DataFrame(
-            {
-                "sample": np.repeat(sample_indices, channel_count),
-                "channel": np.tile(channel_names, values.shape[0]),
-                "value": values.reshape(-1),
-            }
-        )
-
-        if include_time and analog_rate:
-            dataframe.insert(1, "time", dataframe["sample"] / analog_rate)
-
-        return dataframe
+        return build_analog_dataframe(self._load(), self.get_metadata(), include_time)
 
     def export_points(
         self,
@@ -351,7 +148,6 @@ class C3DDataReader:
         Note:
             CSV output is automatically sanitized to prevent Excel Formula Injection.
         """
-
         if not (output_path is not None):
             raise ValueError("output_path must be provided")
         if not (output_path is not None):
@@ -362,8 +158,13 @@ class C3DDataReader:
             residual_nan_threshold=residual_nan_threshold,
             target_units=target_units,
         )
-        return self._export_dataframe(
-            dataframe, output_path, file_format, sanitize=True
+        return export_dataframe(
+            dataframe,
+            output_path,
+            file_format,
+            self.file_path.name,
+            self.get_metadata().units,
+            sanitize=True,
         )
 
     def export_analog(
@@ -387,14 +188,18 @@ class C3DDataReader:
         Note:
             CSV output is automatically sanitized to prevent Excel Formula Injection.
         """
-
         if not (output_path is not None):
             raise ValueError("output_path must be provided")
         if not (output_path is not None):
             raise ValueError("output_path must be provided")
         dataframe = self.analog_dataframe(include_time=include_time)
-        return self._export_dataframe(
-            dataframe, output_path, file_format, sanitize=True
+        return export_dataframe(
+            dataframe,
+            output_path,
+            file_format,
+            self.file_path.name,
+            self.get_metadata().units,
+            sanitize=True,
         )
 
     def get_force_plate_channels(self) -> dict[int, dict[str, str]]:
@@ -410,53 +215,7 @@ class C3DDataReader:
             {1: {'fx': 'Fx1', 'fy': 'Fy1', 'fz': 'Fz1',
                  'mx': 'Mx1', 'my': 'My1', 'mz': 'Mz1'}, ...}
         """
-        metadata = self.get_metadata()
-        labels = metadata.analog_labels
-
-        # Patterns for force plate detection
-        # Standard: Fx1, Fy1, Fz1, Mx1, My1, Mz1
-        # AMTI: FP1_Fx, FP1_Fy, etc.
-        # Kistler: Force.X1, Force.Y1, etc.
-        import re
-
-        plate_channels: dict[int, dict[str, str]] = {}
-
-        # Pattern 1: Standard suffix (Fx1, Fy1, Fz1, Mx1, My1, Mz1)
-        standard_pattern = re.compile(r"^(?:Force\.)?([FfMm])([xyzXYZ])(\d+)$")
-        # Pattern 2: Prefix style (FP1_Fx, FP1_Fy, etc.)
-        prefix_pattern = re.compile(r"^(?:FP|fp)?(\d+)[_.]?([FfMm])([xyzXYZ])$")
-
-        for label in labels:
-            label_stripped = label.strip()
-
-            # Try standard pattern first
-            match = standard_pattern.match(label_stripped)
-            if match:
-                force_or_moment = match.group(1).lower()  # 'f' or 'm'
-                axis = match.group(2).lower()  # 'x', 'y', 'z'
-                plate_num = int(match.group(3))
-
-                if plate_num not in plate_channels:
-                    plate_channels[plate_num] = {}
-
-                key = f"{force_or_moment}{axis}"  # 'fx', 'fy', 'fz', 'mx', 'my', 'mz'
-                plate_channels[plate_num][key] = label
-                continue
-
-            # Try prefix pattern
-            match = prefix_pattern.match(label_stripped)
-            if match:
-                plate_num = int(match.group(1))
-                force_or_moment = match.group(2).lower()
-                axis = match.group(3).lower()
-
-                if plate_num not in plate_channels:
-                    plate_channels[plate_num] = {}
-
-                key = f"{force_or_moment}{axis}"
-                plate_channels[plate_num][key] = label
-
-        return plate_channels
+        return detect_force_plate_channels(self.get_metadata().analog_labels)
 
     def force_plate_dataframe(
         self,
@@ -489,69 +248,28 @@ class C3DDataReader:
         if not (include_time is not None):
             raise ValueError("include_time must be provided")
         plate_channels = self.get_force_plate_channels()
-
-        if not plate_channels:
-            logger.warning(
-                "No force plate channels detected in C3D file. "
-                "Expected channels like Fx1, Fy1, Fz1, Mx1, My1, Mz1."
-            )
-            return pd.DataFrame(
-                columns=self._force_plate_columns(include_time, compute_cop)
-            )
-
-        # Filter to specific plate if requested
-        if plate_number is not None:
-            if plate_number not in plate_channels:
-                raise ValueError(
-                    f"Force plate {plate_number} not found. "
-                    f"Available plates: {list(plate_channels.keys())}"
-                )
-            plate_channels = {plate_number: plate_channels[plate_number]}
-
-        # Get analog data
         analog_df = self.analog_dataframe(include_time=False)
         metadata = self.get_metadata()
-        analog_rate = metadata.analog_rate
-
-        # Pivot to wide format
-        analog_wide = analog_df.pivot(
-            index="sample", columns="channel", values="value"
-        ).reset_index()
-
-        result_dfs = []
-
-        required_keys = {"fx", "fy", "fz", "mx", "my", "mz"}
-
-        for plate_num, channels in sorted(plate_channels.items()):
-            plate_df = self._build_plate_dataframe(
-                plate_num,
-                channels,
-                required_keys,
-                analog_wide,
-                compute_cop,
-                ground_height,
-            )
-            if plate_df is not None:
-                result_dfs.append(plate_df)
-
-        if not result_dfs:
-            return pd.DataFrame(
-                columns=self._force_plate_columns(include_time, compute_cop)
-            )
-
-        result = pd.concat(result_dfs, ignore_index=True)
-
-        if include_time and analog_rate:
-            result.insert(1, "time", result["sample"] / analog_rate)
-
-        logger.info(
-            "Extracted force plate data for %d plates, %d samples from %s",
-            len(plate_channels),
-            len(result),
+        return build_force_plate_dataframe(
+            plate_channels,
+            analog_df,
+            metadata.analog_rate,
             self.file_path.name,
+            plate_number,
+            include_time,
+            compute_cop,
+            ground_height,
         )
 
-        return result
+    def get_force_plate_count(self) -> int:
+        """Return the number of detected force plates."""
+        return len(self.get_force_plate_channels())
+
+    def _load(self) -> C3DMapping:
+        """Load the C3D file if not already loaded."""
+        if self._c3d_data is None:
+            self._c3d_data = load_c3d(self.file_path)
+        return self._c3d_data
 
     @staticmethod
     def _force_plate_columns(
@@ -559,321 +277,4 @@ class C3DDataReader:
         compute_cop: bool,
     ) -> list[str]:
         """Return column names for an empty force plate DataFrame."""
-        if not (include_time is not None):
-            raise ValueError("include_time must be provided")
-        if not (include_time is not None):
-            raise ValueError("include_time must be provided")
-        columns = ["sample", "plate", "fx", "fy", "fz", "mx", "my", "mz"]
-        if include_time:
-            columns.insert(1, "time")
-        if compute_cop:
-            columns.extend(["cop_x", "cop_y", "cop_z"])
-        return columns
-
-    @staticmethod
-    def _build_plate_dataframe(
-        plate_num: int,
-        channels: dict[str, str],
-        required_keys: set[str],
-        analog_wide: pd.DataFrame,
-        compute_cop: bool,
-        ground_height: float,
-    ) -> pd.DataFrame | None:
-        """Build a DataFrame for a single force plate, or None if channels missing."""
-        if not (plate_num is not None):
-            raise ValueError("plate_num must be provided")
-        if not (plate_num is not None):
-            raise ValueError("plate_num must be provided")
-        missing_keys = required_keys - set(channels.keys())
-        if missing_keys:
-            logger.warning(
-                f"Force plate {plate_num} missing channels: {missing_keys}. Skipping."
-            )
-            return None
-
-        plate_df = pd.DataFrame(
-            {
-                "sample": analog_wide["sample"],
-                "plate": plate_num,
-                "fx": analog_wide[channels["fx"]].to_numpy(),
-                "fy": analog_wide[channels["fy"]].to_numpy(),
-                "fz": analog_wide[channels["fz"]].to_numpy(),
-                "mx": analog_wide[channels["mx"]].to_numpy(),
-                "my": analog_wide[channels["my"]].to_numpy(),
-                "mz": analog_wide[channels["mz"]].to_numpy(),
-            }
-        )
-
-        if compute_cop:
-            fz = plate_df["fz"].to_numpy()
-            mx = plate_df["mx"].to_numpy()
-            my = plate_df["my"].to_numpy()
-
-            min_force_threshold = 10.0  # [N] minimum force for valid COP
-            valid_contact = np.abs(fz) > min_force_threshold
-
-            plate_df["cop_x"] = np.where(valid_contact, -my / fz, np.nan)
-            plate_df["cop_y"] = np.where(valid_contact, mx / fz, np.nan)
-            plate_df["cop_z"] = np.where(valid_contact, ground_height, np.nan)
-
-        return plate_df
-
-    def get_force_plate_count(self) -> int:
-        """Return the number of detected force plates."""
-        return len(self.get_force_plate_channels())
-
-    def _get_point_parameters(self) -> dict[str, Any]:
-        """Get POINT parameters from the C3D file."""
-        c3d_data = self._load()
-        try:
-            return cast(dict[str, Any], c3d_data["parameters"]["POINT"])
-        except KeyError as error:  # pragma: no cover - defensive guard
-            raise ValueError(
-                f"POINT parameters missing from C3D file: {self.file_path}"
-            ) from error
-
-    def _get_analog_parameters(self) -> dict[str, Any] | None:
-        """Get ANALOG parameters from the C3D file, if present."""
-        c3d_data = self._load()
-        analog_params = c3d_data["parameters"].get("ANALOG")
-        return (
-            cast(dict[str, Any], analog_params) if analog_params is not None else None
-        )
-
-    def _get_analog_details(self) -> tuple[list[str], float | None, list[str]]:
-        """Get analog channel labels, sample rate, and units from the C3D file."""
-        analog_parameters = self._get_analog_parameters()
-        analog_array = self._load()["data"]["analogs"]
-        channel_count = analog_array.shape[1]
-
-        if analog_parameters is None:
-            labels = []
-            units = []
-            analog_rate = None
-        else:
-            labels = [
-                label.strip()
-                for label in analog_parameters.get("LABELS", {}).get("value", [])
-            ]
-            units = [
-                unit.strip()
-                for unit in analog_parameters.get("UNITS", {}).get("value", [])
-            ]
-            analog_rate = float(analog_parameters.get("RATE", {}).get("value", [0])[0])
-
-        if not labels and channel_count > 0:
-            labels = [f"Analog_{idx + 1}" for idx in range(channel_count)]
-
-        # Ensure units list checks out
-        if len(units) < len(labels):
-            units.extend([""] * (len(labels) - len(units)))
-        elif len(units) > len(labels):
-            units = units[: len(labels)]
-
-        return labels, analog_rate, units
-
-    def _get_events(self) -> list[C3DEvent]:
-        """Extract event markers from the C3D file."""
-        c3d_data = self._load()
-        event_parameters = c3d_data["parameters"].get("EVENT")
-        if not event_parameters:
-            return []
-
-        labels_raw: Iterable[str] = event_parameters.get("LABELS", {}).get("value", [])
-        times = event_parameters.get("TIMES", {}).get("value")
-        if times is None:
-            return []
-
-        times_array = np.asarray(times)
-        if times_array.ndim == 2:
-            times_array = times_array[1, :]
-
-        events: list[C3DEvent] = []
-        for idx, label in enumerate(labels_raw):
-            time_value = float(times_array[idx]) if idx < len(times_array) else np.nan
-            if np.isfinite(time_value):
-                events.append(C3DEvent(label=str(label).strip(), time=time_value))
-
-        return events
-
-    def _load(self) -> C3DMapping:
-        """Load the C3D file if not already loaded."""
-        if self._c3d_data is None:
-            if ezc3d is None:
-                raise ImportError(
-                    "ezc3d is required for C3D file reading. "
-                    "Install it with: pip install ezc3d\n"
-                    "Note: ezc3d requires Python >=3.10. "
-                    "For Python 3.9, this functionality is not available."
-                )
-            if not self.file_path.exists():
-                raise FileNotFoundError(f"File not found: {self.file_path}")
-            self._c3d_data = ezc3d.c3d(str(self.file_path))
-        return self._c3d_data
-
-    @staticmethod
-    def _sanitize_for_csv(value: Any) -> Any:
-        """Sanitize a value to prevent CSV injection."""
-        if not isinstance(value, str):
-            return value
-        if value.startswith(("=", "+", "-", "@")):
-            return f"'{value}"
-        return value
-
-    @staticmethod
-    def _unit_scale(current_units: str, target_units: str | None) -> float:
-        """Calculate scaling factor for unit conversion."""
-        if target_units is None:
-            return 1.0
-
-        normalized_current = current_units.lower()
-        normalized_target = target_units.lower()
-
-        if normalized_current == normalized_target:
-            return 1.0
-
-        # Define conversion factors to meters
-        to_meters = {
-            "m": 1.0,
-            "mm": 0.001,
-            "mm^2": 0.000001,  # Added minimal robust area unit support for consistency
-            "cm": 0.01,
-            "in": 0.0254,
-            "ft": 0.3048,
-        }
-        # Note: Original code only had length units.
-        # Stick to original:
-        to_meters = {
-            "m": 1.0,
-            "mm": 0.001,
-            "cm": 0.01,
-            "in": 0.0254,
-            "ft": 0.3048,
-        }
-
-        if normalized_current not in to_meters:
-            raise ValueError(f"Unsupported source unit: {current_units}")
-        if normalized_target not in to_meters:
-            raise ValueError(f"Unsupported target unit: {target_units}")
-
-        return to_meters[normalized_current] / to_meters[normalized_target]
-
-    def _export_dataframe(
-        self,
-        dataframe: pd.DataFrame,
-        output_path: Path | str,
-        file_format: str | None,
-        sanitize: bool = True,
-    ) -> Path:
-        """Export a DataFrame to CSV, JSON, or NPZ format.
-
-        Includes validation, versioning, and telemetry.
-        """
-        if not (dataframe is not None):
-            raise ValueError("dataframe must be provided")
-        if not (dataframe is not None):
-            raise ValueError("dataframe must be provided")
-        path = Path(output_path).resolve()
-
-        self._validate_export_path(path)
-
-        if not file_format:
-            if not path.suffix:
-                raise ValueError(
-                    "File format could not be inferred from the path suffix."
-                )
-            file_format = path.suffix.lstrip(".")
-
-        normalized_format = file_format.lower()
-        path.parent.mkdir(parents=True, exist_ok=True)
-
-        with log_execution_time(f"export_{normalized_format}"):
-            metadata = {
-                "schema_version": SCHEMA_VERSION,
-                "created_at_utc": datetime.now(timezone.utc).isoformat(),  # noqa: UP017
-                "source_file": self.file_path.name,
-                "row_count": len(dataframe),
-                "units": self.get_metadata().units,
-            }
-            self._write_export(
-                path,
-                normalized_format,
-                dataframe,
-                metadata,
-                sanitize,
-            )
-
-        return path
-
-    @staticmethod
-    def _validate_export_path(path: Path) -> None:
-        """Validate export path for security (prevent directory traversal)."""
-        base_dir = Path.cwd().resolve()
-
-        import inspect
-
-        frame = inspect.currentframe()
-        is_security_test = False
-        try:
-            while frame:
-                if frame.f_code.co_name == "test_security_prevents_directory_traversal":
-                    is_security_test = True
-                    break
-                frame = frame.f_back
-        finally:
-            del frame
-
-        is_test_env = not is_security_test and any(
-            [
-                "pytest" in str(base_dir),
-                "test" in str(base_dir).lower(),
-                "/tmp/pytest" in str(path),  # nosec B108 - string comparison only, not creating a temp file
-                "pytest" in str(path),
-            ]
-        )
-
-        if not is_test_env and base_dir not in path.parents and path != base_dir:
-            raise ValueError(
-                f"Security: Refusing to output to {path} "
-                f"(outside project root {base_dir})"
-            )
-
-    def _write_export(
-        self,
-        path: Path,
-        fmt: str,
-        dataframe: pd.DataFrame,
-        metadata: dict[str, Any],
-        sanitize: bool,
-    ) -> None:
-        """Write dataframe to disk in the given format."""
-        if fmt == "csv":
-            df_to_export = dataframe.copy() if sanitize else dataframe
-            if sanitize:
-                for col in df_to_export.select_dtypes(
-                    include=[object, "string"]
-                ).columns:
-                    df_to_export[col] = df_to_export[col].apply(self._sanitize_for_csv)
-            df_to_export.to_csv(path, index=False)
-
-            meta_path = path.with_name(f"{path.stem}_meta.json")
-            with open(meta_path, "w") as f:
-                json.dump(metadata, f, indent=2)
-
-        elif fmt == "json":
-            output = {
-                "metadata": metadata,
-                "data": dataframe.to_dict(orient="records"),
-            }
-            with open(path, "w") as f:
-                json.dump(output, f, indent=2)
-
-        elif fmt == "npz":
-            arrays = {column: dataframe[column].to_numpy() for column in dataframe}
-            np.savez(path, _metadata=json.dumps(metadata), **arrays)
-
-        else:
-            raise ValueError(
-                f"Unsupported export format: '{fmt}'. "
-                "Supported formats: csv, json, npz."
-            )
+        return force_plate_columns(include_time, compute_cop)
